% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer_design_functions.R
\name{validate_primers}
\alias{validate_primers}
\title{Primer design}
\usage{
validate_primers(primers, seqs)
}
\arguments{
\item{primers}{A data.frame as produced by design_primers()}

\item{seqs}{Input sequences (DNAStringSet object) that primers should be compared to.}
}
\value{
data.frame as from design_primers() with additional information on mismatching sequences.
}
\description{
This function searches for mismatches of primer-probe sets with a set of sequences.
}
\details{
This function searches for sequences in the data set that have mismatches with any of the primer-probe sets and stores the sequence name and minimum alignment score of any mismatching sequence in the primers data.frame.
}
\keyword{primer}
